requirements.txt
pygments>=2.4.0
regex>=2016.3.2
python-dateutil>=2.6.1
ruamel.yaml>=0.16.10
dateparser>=0.6.0
ipython>=6.2.0
lxml>=3.5.0
csvtomd>=0.3.0
pyyaml>=5.0.1
boto3>1.1.11
# Tests
nose>=1.3.0
pandas>=0.22.0
openpyxl>=2.5.4
pylint>=1.8.2
pytest
pytest-flake8
scikit-learn>0.22.0
Flask==1.1.2
flask-cors
pyarrow==0.17.0
